Revert "Revert "bismiilah testimonial""
This reverts commit bc3211ef549ea2566f9c6c3047cab1a3deef00ed.

+++ database/migrations/2022_03_09_141937_create_portofolio_tags_table.php
@@ -15,7 +15,7 @@ class CreatePortofolioTagsTable extends Migration
     {
         Schema::create('portofolio_tags', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('portofolio_id');
+            $table->unsignedBigInteger('portofolio_id');
             $table->string('tag');
             $table->timestamps();
             $table->foreign('portofolio_id')->references('id')->on('portofolios')->onUpdate('cascade')->onDelete('cascade');
